# More efficient parsing of numbers