--- a/zio-json/shared/src/test/scala/zio/json/CodecSpec.scala
+++ b/zio-json/shared/src/test/scala/zio/json/CodecSpec.scala
@@ -35,7 +35,6 @@ object CodecSpec extends ZIOSpecDefault {
           )
         },
         test("primitives") {
-          val exampleBDString = "234234.234"
           // this big integer consumes more than 256 bits
           assert(
             "170141183460469231731687303715884105728489465165484668486513574864654818964653168465316546851"