Below are throughout results with Scala 3 and JDK-21 on Intel® Core™ i7-11800H:

#### Before
```txt
Benchmark                       (size)   Mode  Cnt        Score       Error  Units
ArrayOfBytesReading.zioJson        128  thrpt    5  1004751.030 ±  7254.736  ops/s
ArrayOfIntsReading.zioJson         128  thrpt    5   651846.166 ±  6159.136  ops/s
ArrayOfLongsReading.zioJson        128  thrpt    5   432257.664 ± 27319.590  ops/s
ArrayOfShortsReading.zioJson       128  thrpt    5   908516.530 ± 25140.642  ops/s
```

#### After
```txt
Benchmark                       (size)   Mode  Cnt        Score       Error  Units
ArrayOfBytesReading.zioJson        128  thrpt    5  1021454.947 ± 22940.718  ops/s
ArrayOfIntsReading.zioJson         128  thrpt    5   691045.897 ± 17081.958  ops/s
ArrayOfLongsReading.zioJson        128  thrpt    5   485679.414 ±  7556.651  ops/s
ArrayOfShortsReading.zioJson       128  thrpt    5   947579.519 ± 11212.803  ops/s
```

Up to 20% speedup was measured with Scala.js